fix(type): Make badge NSNumber instead of NSInteger
* This let's us differentiate between null vs 0. The payload from APNS will be already be an NSNumber: https://developer.apple.com/documentation/usernotifications/unnotificationcontent/badge?language=objc
* This fixes a bug when a notification had "SetTo" = 0, our check for the existence of notification.badge could not differentiate null vs 0, and we detected 0 as lack of existence.
* Note: we will still cache and read badge count (ONESIGNAL_BADGE_KEY) as NSInteger since that is what we have been doing, and avoid having to change types / migration.

+++ iOS_SDK/OneSignalSDK/OneSignalExtension/OneSignalExtensionBadgeHandler.m
@@ -35,7 +35,7 @@
     //make sure the OneSignal cached value is updated to this value
     if (!notification.badgeIncrement) {
         if (notification.badge)
-            [OneSignalExtensionBadgeHandler updateCachedBadgeValue:notification.badge];
+            [OneSignalExtensionBadgeHandler updateCachedBadgeValue:notification.badge.intValue];
         
         return;
     }
